Improve helpers structure
