+++ api/src/database/helpers/geometry/types.ts
@@ -1,0 +1,52 @@
+import { Knex } from 'knex';
+import { stringify as geojsonToWKT, GeoJSONGeometry } from 'wellknown';
+import { Field, RawField } from '@directus/shared/types';
+
+export abstract class GeometryHelper {
+	constructor(protected knex: Knex) {}
+	supported(): boolean | Promise<boolean> {
+		return true;
+	}
+	isTrue(expression: Knex.Raw) {
+		return expression;
+	}
+	isFalse(expression: Knex.Raw) {
+		return expression.wrap('NOT ', '');
+	}
+	createColumn(table: Knex.CreateTableBuilder, field: RawField | Field) {
+		const type = field.type.split('.')[1] ?? 'geometry';
+		return table.specificType(field.field, type);
+	}
+	asText(table: string, column: string): Knex.Raw {
+		return this.knex.raw('st_astext(??.??) as ??', [table, column, column]);
+	}
+	fromText(text: string): Knex.Raw {
+		return this.knex.raw('st_geomfromtext(?, 4326)', text);
+	}
+	fromGeoJSON(geojson: GeoJSONGeometry): Knex.Raw {
+		return this.fromText(geojsonToWKT(geojson));
+	}
+	_intersects(key: string, geojson: GeoJSONGeometry): Knex.Raw {
+		const geometry = this.fromGeoJSON(geojson);
+		return this.knex.raw('st_intersects(??, ?)', [key, geometry]);
+	}
+	intersects(key: string, geojson: GeoJSONGeometry): Knex.Raw {
+		return this.isTrue(this._intersects(key, geojson));
+	}
+	nintersects(key: string, geojson: GeoJSONGeometry): Knex.Raw {
+		return this.isFalse(this._intersects(key, geojson));
+	}
+	_intersects_bbox(key: string, geojson: GeoJSONGeometry): Knex.Raw {
+		const geometry = this.fromGeoJSON(geojson);
+		return this.knex.raw('st_intersects(??, ?)', [key, geometry]);
+	}
+	intersects_bbox(key: string, geojson: GeoJSONGeometry): Knex.Raw {
+		return this.isTrue(this._intersects_bbox(key, geojson));
+	}
+	nintersects_bbox(key: string, geojson: GeoJSONGeometry): Knex.Raw {
+		return this.isFalse(this._intersects_bbox(key, geojson));
+	}
+	collect(table: string, column: string): Knex.Raw {
+		return this.knex.raw('st_astext(st_collect(??.??))', [table, column]);
+	}
+}
